requirements.txt
fastapi[standard]
sqlalchemy
jinja2
python-multipart
pyzbar
Pillow
passlib
isbnlib
